+++ typescript/deploy/src/config.ts
@@ -3,7 +3,7 @@ import { ethers } from 'ethers';
 import { ChainMap, ChainName } from '@abacus-network/sdk';
 
 export interface CheckerViolation {
-  network: ChainName;
+  chain: ChainName;
   type: string;
   expected: any;
   actual: any;
@@ -18,7 +18,7 @@ export type TransactionConfig = {
   signer?: ethers.Signer;
 };
 
-export type EnvironmentConfig<Networks extends ChainName> = ChainMap<
-  Networks,
+export type EnvironmentConfig<Chain extends ChainName> = ChainMap<
+  Chain,
   TransactionConfig
 >;