+++ typescript/infra/config/environments/dev/core.ts
@@ -1,9 +1,9 @@
 import { CoreConfig } from '@abacus-network/deploy';
 import { ChainMap } from '@abacus-network/sdk';
 
-import { DevNetworks } from './domains';
+import { DevChains } from './chains';
 
-export const core: ChainMap<DevNetworks, CoreConfig> = {
+export const core: ChainMap<DevChains, CoreConfig> = {
   alfajores: {
     validatorManager: {
       validators: [