+++ typescript/infra/config/environments/test/index.ts
@@ -3,12 +3,12 @@ import { utils } from '@abacus-network/deploy';
 import { CoreEnvironmentConfig } from '../../../src/config';
 
 import { agent } from './agent';
+import { TestChains, testConfigs } from './chains';
 import { controller } from './controller';
 import { core } from './core';
-import { TestNetworks, testConfigs } from './domains';
 import { infra } from './infra';
 
-export const environment: CoreEnvironmentConfig<TestNetworks> = {
+export const environment: CoreEnvironmentConfig<TestChains> = {
   transactionConfigs: testConfigs,
   agent,
   core,